🐛fix invalids check
A problem occured when running the diff script, checking for invalid
suffixes. The if statement had a typo that did not triggger on older
versions of bash. The bug was not catched in local development, because
the machine was using bash 3.2, but in environments running different
versions it was discovered.